requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
twilio==6.34.0
urllib3==1.25.7
python-dotenv==0.10.3
